+++ wled00/cfg.cpp
@@ -116,7 +116,6 @@ bool deserializeConfig(JsonObject doc, bool fromFS) {
       bool reversed = elm["rev"];
       bool refresh = elm["ref"] | false;
       ledType |= refresh << 7; // hack bit 7 to indicate strip requires off refresh
-      s++;
       if (fromFS) {
         BusConfig bc = BusConfig(ledType, pins, start, length, colorOrder, reversed, skipFirst);
         mem += BusManager::memUsage(bc);
@@ -126,6 +125,7 @@ bool deserializeConfig(JsonObject doc, bool fromFS) {
         busConfigs[s] = new BusConfig(ledType, pins, start, length, colorOrder, reversed, skipFirst);
         doInitBusses = true;
       }
+      s++;
     }
     // finalization done in beginStrip()
   }
@@ -455,7 +455,9 @@ bool deserializeConfig(JsonObject doc, bool fromFS) {
   }
 
   if (fromFS) return needsSave;
+  // if from /json/cfg
   doReboot = doc[F("rb")] | doReboot;
+  if (doInitBusses) return false; // no save needed, will do after bus init in wled.cpp loop
   return (doc["sv"] | true);
 }